requirements.txt
# ============================================
# DEAR X PROXY BOT - COMPLETE DEPENDENCIES
# ============================================

# Telegram Bot Framework
python-telegram-bot==20.7
telegram==0.0.1

# Async HTTP Client & Web Scraping
aiohttp==3.9.1
aiofiles==23.2.1
requests==2.31.0
httpx==0.25.2

# Proxy Protocol Support
PySocks==1.7.1
aiohttp-socks==0.8.4
python-socks==2.4.3

# HTML/XML Parsing
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Environment Variables
python-dotenv==1.0.0

# File Handling & Path Management
pathlib2==2.3.7
aiofiles==23.2.1

# Date/Time Utilities
python-dateutil==2.8.2

# Logging & Monitoring
colorlog==6.8.0
loguru==0.7.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# JSON & Data Serialization
orjson==3.9.10
ujson==5.9.0

# Regular Expressions (advanced)
regex==2023.10.3

# Async Utilities
asyncio-throttle==1.0.2
aiodns==3.1.1
cchardet==2.1.7

# Network Utilities
dnspython==2.4.2
urllib3==2.1.0

# SSL/TLS
certifi==2023.11.17
pyOpenSSL==23.3.0

# Database
# (SQLite3 is built-in with Python)

# Text Processing
chardet==5.2.0
charset-normalizer==3.3.2

# Progress Bars (for CLI testing)
tqdm==4.66.1
alive-progress==3.1.5

# Error Tracking
sentry-sdk==1.38.0

# Rate Limiting
ratelimit==2.2.1
limits==3.7.0

# Caching
cachetools==5.3.2

# Type Checking & Validation
pydantic==2.5.2
typing-extensions==4.9.0

# Testing (optional but recommended)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality (optional)
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Security
cryptography==41.0.7

# IP Address Utilities
ipaddress==1.0.23
netaddr==0.10.1

# User Agent Rotation
fake-useragent==1.4.0

# Async DNS Resolution
aiodns==3.1.1

# HTTP/2 Support
h2==4.1.0
hyper==0.7.0

# WebSocket Support (if needed)
websockets==12.0

# HTML Parsing Alternatives
pyquery==2.0.0
selectolax==0.3.17

# CSV Processing
csv23==0.3.3

# File Type Detection
python-magic==0.4.27

# Compression
zstandard==0.22.0

# Unicode Handling
unidecode==1.3.7

# URL Processing
furl==2.1.3
yarl==1.9.4

# Async Context Management
async-timeout==4.0.3

# Multi-threading Support
concurrent-futures==3.1.1


# System Information
psutil==5.9.6

# Performance Monitoring
memory-profiler==0.61.0

# HTTP Request Retrying
tenacity==8.2.3
backoff==2.2.1

# API Rate Limiting
pyrate-limiter==3.1.1

# Time Zone Handling
pytz==2023.3.post1
tzdata==2023.3

# Proxy Detection & Validation
validators==0.22.0

# String Similarity (for proxy deduplication)
python-levenshtein==0.23.0
fuzzywuzzy==0.18.0

# Process Management
supervisor==4.2.5

# Configuration Management
configparser==6.0.0
pyyaml==6.0.1
toml==0.10.2

# CLI Enhancements
click==8.1.7
rich==13.7.0

# Async Iterators
aioitertools==0.11.0

# HTTP Sessions
requests-toolbelt==1.0.0

# Proxy Rotation
iterproxy==0.1

# Geolocation (optional - for IP analysis)
geoip2==4.7.0
maxminddb==2.5.1

# ============================================
# MINIMAL REQUIREMENTS (If size matters)
# ============================================
# If you want minimal installation, use only:
# python-telegram-bot==20.7
# aiohttp==3.9.1
# aiohttp-socks==0.8.4
# python-dotenv==1.0.0
# beautifulsoup4==4.12.2
# requests==2.31.0
# PySocks==1.7.1

# ============================================
# INSTALLATION NOTES
# ============================================
# 
# Standard Installation:
# pip install -r requirements.txt
#
# Minimal Installation:
# pip install python-telegram-bot aiohttp aiohttp-socks python-dotenv
#
# With Testing:
# pip install -r requirements.txt
# pip install pytest pytest-asyncio
#
# For Development:
# pip install -r requirements.txt
# pip install black flake8 mypy
#
# ============================================
